fixed efficiency test names and added inv_sym test (#2596)
